reports/M4/webapp/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.1+cpu
torch==1.8.1+cpu 
streamlit
Pillow